Rename index.js to api.js since it is not the bundled file
